+++ resources/views/comment-installer.blade.php
@@ -9,7 +9,32 @@
         <h1>Comment installer un script</h1>
 
         <p> 
-            TODO : explication en fonction du naviguateur.
+            Prérequis : <br>
+        </p>
+        <p>
+            Pour Chrome, installer <a target="_blank" href="https://chrome.google.com/webstore/detail/tampermonkey/dhdgffkkebhmkfjojejmpbldmpobfkfo?hl=fr">TamperMonkey</a> <br>
+        </p>
+        <p>
+            Pour Firefox, installer <a target="_blank" href="https://addons.mozilla.org/fr/firefox/addon/greasemonkey/">GreaseMonkey</a>
+        </p>
+        
+        <p>
+            Ensuite, cliquez sur le bouton "Installer" des scripts
+        </p>
+
+        <h1>Comment installer un skin</h1>
+
+        <p> 
+            Prérequis, installer Stylish : 
+        </p>
+        <p>
+            Pour <a target="_blank" href="https://chrome.google.com/webstore/detail/stylish-custom-themes-for/fjnbnpbmkenffdnngjfgmeleoegfcffe?hl=fr">Chrome </a> 
+        </p>
+        <p>
+            Pour  <a target="_blank" href="https://addons.mozilla.org/fr/firefox/addon/stylish/">Firefox</a>
+        </p>
+         <p>
+            Ensuite, cliquez sur le bouton "Installer" des styles puis sur "Install with Stylish"
         </p>
 
     </div>