+++ resources/views/mails/contact.blade.php
@@ -1,0 +1,10 @@
+Nouveau message du formulaire de contact de jvscript.io <br>
+
+<p>
+    Sender : {{$email}}
+</p>
+
+
+<p>
+    {{$message_body}}
+</p>