+++ routes/web.php
@@ -12,35 +12,42 @@
  */
 
 Route::get('/', 'JvscriptController@index')->name('home');
+ 
 
-Route::get('/comment-installer', function () {
-    return view('comment-installer');
-});
+//forms
+Route::get('/script/ajout', 'JvscriptController@formScript')->name('script.form');
+Route::get('/skin/ajout', 'JvscriptController@formSkin')->name('skin.form');
 
-Route::get('/ajout', function () {
-    return view('script.ajout');
-})->name("ajout-form");
+//form action (store in db)
+Route::post('/script/ajout', 'JvscriptController@storeScript')->name('script.store');
+Route::post('/skin/ajout', 'JvscriptController@storeSkin')->name('skin.store');
 
-Route::post('/ajout', 'JvscriptController@storeScript')->name('script.store');
 
 Route::get('/script/{slug}', 'JvscriptController@showScript')->name('script.show');
-
 Route::get('/script/install/{slug}', 'JvscriptController@installScript')->name('script.install');
-
 Route::get('/script/note/{slug}/{note}', 'JvscriptController@noteScript')->name('script.note');
 
-Route::get('/contact', function () {
-    return view('contact');
-});
 
+Route::get('/skin/{slug}', 'JvscriptController@showSkin')->name('skin.show');
+Route::get('/skin/install/{slug}', 'JvscriptController@installSkin')->name('skin.install');
+Route::get('/skin/note/{slug}/{note}', 'JvscriptController@noteSkin')->name('skin.note');
+
+
+//contact form
+Route::get('/contact/{message_body?}', function ($message_body = null) {
+    return view('contact', ['message_body' => $message_body]);
+})->name('contact.form');
+//contact action
+Route::post('/contact', 'JvscriptController@contactSend')->name('contact.send');
+
+
+//static views 
 Route::get('/developpeurs', function () {
     return view('developpeurs');
 });
-
-
-
+Route::get('/aide', function () {
+    return view('comment-installer');
+})->name('aide');
 
 
 //Auth::routes();
-
-//Route::get('/home', 'HomeController@index');