+++ changelog.md
@@ -3,4 +3,4 @@
 
 ## v0.0.1 - 2017-01-20
 
-- Fonctionnalités des scripts (ajout, listing) 
+- Fonctionnalités des scripts/skins